requirements.txt
aiocontextvars==0.2.2
aria2p==0.8.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
console-progressbar==1.1.2
contextvars==2.4
cryptography==2.8
decorator==4.4.2
future==0.18.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
immutables==0.11
loguru==0.4.1
oauth2client==4.1.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDrive==1.3.1
python-telegram-bot==12.5
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
six==1.14.0
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.8
websocket-client==0.57.0
gunicorn
